Use interface insted of module
