+++ mimic-cross.deno/apt/package.ts
@@ -1,0 +1,10 @@
+export interface PackageInfo {
+  postInstall?: string;
+  isCrossTool?: boolean;
+  blockList?: string[];
+}
+
+export interface PackageRecipe {
+  nameResolver?(name: string, info: PackageInfo): string[];
+  postInstall?(name: string, info: PackageInfo): Promise<void>;
+}
