--- mimic-cross.deno/apt/package_info.ts
@@ -1,5 +1,0 @@
-export interface PackageInfo {
-  postInstall?: string;
-  isCrossTool?: boolean;
-  blockList?: string[];
-}
